+++ packages/hero_ui/lib/src/components/components.dart
@@ -2,4 +2,5 @@
 
 export 'button/button.dart';
 export 'button_group/button_group.dart';
+export 'close_button/close_button.dart';
 export 'spinner/spinner.dart';